+++ CHANGELOG.md
@@ -11,6 +11,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+## [0.3.0] - 2026-05-28
+
+### Added
+
+### Changed
+
 - Changed persistent panel and form-dialog errors to render as copyable red alert boxes.
 
 ### Deprecated